requirement.txt
asgiref==3.2.9
autopep8==1.5.3
certifi==2020.4.5.2
chardet==3.0.4
Django==3.0.7
django-environ==0.4.5
djangorestframework==3.11.0
feedparser==5.2.1
idna==2.9
psycopg2==2.8.5
pycodestyle==2.6.0
pytz==2020.1
requests==2.24.0
sqlparse==0.3.1
toml==0.10.1
urllib3==1.25.9
